+++ CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.3.37] = 2020-10-29
+## [0.3.36] = 2020-10-29
 - allow the language to be defined in the URL
+
+## [0.3.35] = 2020-10-29
+- fix applet naming for duplicated applets
+
+## [0.3.34] = 2020-10-29
+- show reviewer dashboard as user's timezone
 
 ## [0.3.33] = 2020-10-23
 - insert option to show/hide version bars